More robust determination of executable location
There are several problems with using the 0-th command-line argument, including
not being accurate when run via the PATH and complications when it is not an absolute
path.

Fortunately, Foundation provides Bundle.main.executableURL. However, at present this
is nil on Linux, so an alternative, working Linux implementation is provided under
a conditional in this commit.

Add support for loading relative to project dir

Add unit tests for relative loading

Support executable location under XCTest runner

Add relative load test in external program